处理消息问题

+++ network/IMessage.go
@@ -1,6 +1,6 @@
 package network
 
 type IMessage interface {
-	GetMsgId() int32 //获取消息ID
+	GetMsgId() int64 //获取消息ID
 	GetData() []byte //获取消息内容
 }
